X11: Add OverrideRedirect attribute to new windows
This flag allows windows to avoid capture by tiling window managers such
as i3. Following this change, windows created with a minimum and maximum
size will not be tiled by the WM when first created. Note that the logic
for specifying the size hints of the window to prevent scaling was
already present. This change just adds the OverrideRedirect flag which
allow them to take effect.