Hook up a runloop observer and a timer for the runloop coroutine
This puts an upper bound on how long the coroutine can execute without
yielding back to the main thread. The runloop observer fires before going to
sleep, which is useful if `forward_event_to_cocoa()` has triggered an inner
event loop.

The timer could be used to wake the event loop from a different thread, but
in practice that has a few milliseconds of latency, and it also requires
finding a reference to the timer from `Runloop::wake()`. Instead, the timer
is configured to fire every few milliseconds all the time, again in the
pursuit of lower latency.