[macOS] Get the NSView for a window with NSWindow::contentView
Instead of holding a reference to the original `NSView` in `Window`, only hold the `NSWindow` and get its view by
calling `NSWindow::contentView`. This allows the user to set a new `contentView` without breaking winit's resize
handling. My specific use case is to replace the `contentView` with an `MTKView` allowing me to render with Metal.